Digital Storytelling Through the SVG ViewBox
The web has historically been a one-dimensional medium: start at the top and scroll down. Unexplored capabilities in the browser let us break out of the 1D world and into the 3D universe.

In this talk, we'll examine animating the SVG viewBox, giving us access to powerful three-dimensional motion effects. We'll use these techniques to provide spatial animations that are high-resolution, performant, and best of all, easily accessible to designers who don't need to learn how to program to deliver cinematic-quality animations on the web.